fix: #423 双轴评审修复——MSGEFFECT 复查补租约(阻断项)+租约计数封装+脚本健壮性
- 阻断:MSGEFFECT fling 等待(可挂 2s)后复查点漏验租约,episode 在窗内启动
  (点按不置 isScrollInProgress)则带锚穿越重开 ±H 互搏——补 hasActiveTransactions
  与 A9 横幅双检同款(评审旁证:其余全部锚定路径等待后均复查租约)
- 封装:activeCount 私有化 backing 只读暴露(计数不变量由 withEpisode 独占);
  KDoc 补线程语义(主线程单写者,调度器保证)
- CardExpandReveal withEpisode 闭合括号配对标记(热文件零缩进 churn)
- 脚本:install-dev.sh 探活早退+mktemp 独立目录;capture.sh 计数改 find|wc
- 复验:四测试类 --rerun 全绿;真机三连点烟测干净收官零战争线

+++ app/src/test/kotlin/dev/leonardo/ocbeacon/ui/screens/chat/scroll/PreRenderCoordinatorTest.kt
@@ -31,11 +31,11 @@ class PreRenderCoordinatorTest {
     fun `nested episodes are counted`() = runTest {
         PreRenderCoordinator.withEpisode {
             PreRenderCoordinator.withEpisode {
-                assertEquals(2, PreRenderCoordinator.activeCount.intValue)
+                assertEquals(2, PreRenderCoordinator.activeCount)
             }
-            assertEquals(1, PreRenderCoordinator.activeCount.intValue)
+            assertEquals(1, PreRenderCoordinator.activeCount)
         }
-        assertEquals(0, PreRenderCoordinator.activeCount.intValue)
+        assertEquals(0, PreRenderCoordinator.activeCount)
     }
 
     @Test
